CON-1082 - moved all reviews related data to reviews collection
