+++ product/constants/index.js
@@ -44,11 +44,6 @@ export const ERROR_PRODUCT_IMAGES = 'ERROR_PRODUCT_IMAGES';
 // CURRENT PRODUCT
 export const RESET_CURRENT_PRODUCT = 'RESET_CURRENT_PRODUCT';
 
-// PRODUCT REVIEWS
-export const REQUEST_PRODUCT_REVIEWS = 'REQUEST_PRODUCT_REVIEWS';
-export const RECEIVE_PRODUCT_REVIEWS = 'RECEIVE_PRODUCT_REVIEWS';
-export const ERROR_PRODUCT_REVIEWS = 'ERROR_PRODUCT_REVIEWS';
-
 export const SET_PRODUCT_ID = 'SET_PRODUCT_ID';
 export const SET_PRODUCT_VARIANT_ID = 'SET_PRODUCT_VARIANT_ID';
 export const SET_PRODUCT_QUANTITY = 'SET_PRODUCT_QUANTITY';
