+++ product/reducers/index.js
@@ -15,7 +15,6 @@ import shippingByProductId from './shippingByProductId';
 import variantsByProductId from './variantsByProductId';
 import optionsByProductId from './optionsByProductId';
 import resultsByHash from './resultsByHash';
-import reviewsByProductId from './reviewsByProductId';
 
 export default combineReducers({
   currentProduct,
@@ -25,7 +24,6 @@ export default combineReducers({
   productsById,
   propertiesByProductId,
   resultsByHash,
-  reviewsByProductId,
   shippingByProductId,
   variantsByProductId,
 });
