+++ reviews/action-creators/receiveProductReviews.js
@@ -1,8 +1,9 @@
-/**
- * Copyright (c) 2017, Shopgate, Inc. All rights reserved.
+/*
+ *  Copyright (c) 2017, Shopgate, Inc. All rights reserved.
  *
- * This source code is licensed under the Apache 2.0 license found in the
- * LICENSE file in the root directory of this source tree.
+ *  This source code is licensed under the Apache 2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ *
  */
 
 import { RECEIVE_PRODUCT_REVIEWS } from '../constants';
