+++ reviews/constants/index.js
@@ -10,3 +10,8 @@ export const REVIEWS_LIFETIME = 900000; // 15 minutes
 export const REQUEST_REVIEWS = 'REQUEST_REVIEWS';
 export const RECEIVE_REVIEWS = 'RECEIVE_REVIEWS';
 export const ERROR_REVIEWS = 'ERROR_REVIEWS';
+
+// PRODUCT REVIEWS
+export const REQUEST_PRODUCT_REVIEWS = 'REQUEST_PRODUCT_REVIEWS';
+export const RECEIVE_PRODUCT_REVIEWS = 'RECEIVE_PRODUCT_REVIEWS';
+export const ERROR_PRODUCT_REVIEWS = 'ERROR_PRODUCT_REVIEWS';
