+++ reviews/reducers/index.js
@@ -8,8 +8,10 @@
 import { combineReducers } from 'redux';
 import reviewsByHash from './reviewsByHash';
 import reviewsById from './reviewsById';
+import reviewsByProductId from './reviewsByProductId';
 
 export default combineReducers({
   reviewsByHash,
   reviewsById,
+  reviewsByProductId,
 });
