+++ reviews/reducers/reviewsByProductId.js
@@ -1,3 +1,11 @@
+/*
+ *  Copyright (c) 2017, Shopgate, Inc. All rights reserved.
+ *
+ *  This source code is licensed under the Apache 2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ *
+ */
+
 import {
   REQUEST_PRODUCT_REVIEWS,
   RECEIVE_PRODUCT_REVIEWS,
@@ -28,7 +36,7 @@ export default function reviewsByProductId(state = {}, action) {
         [action.productId]: {
           ...state[action.productId],
           isFetching: false,
-          reviews: action.reviews,
+          reviews: action.reviews.map(review => review.id),
           totalReviewCount: action.totalReviewCount,
         },
       };
